requirements.txt
kafka-python
requests
beautifulsoup4
flask
